requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
decorator==5.2.1
essentia==2.1b6.dev1389
fastapi==0.135.1
h11==0.16.0
idna==3.11
internetarchive==5.8.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
msgpack==1.1.2
numba==0.64.0
numpy==2.4.2
packaging==26.0
platformdirs==4.9.4
pooch==1.9.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
